:recycle: Move AccountVisual to toolbox
