+++ src/components/send/form/bookmarkAutoSuggest.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { getNetworkCode } from '../../../utils/api/btc/network';
 import { validateAddress } from '../../../utils/validators';
-import AccountVisual from '../../accountVisual';
+import AccountVisual from '../../toolbox/accountVisual';
 import AutoSuggest from '../../toolbox/autoSuggest';
 import styles from './form.css';
