+++ src/components/toolbox/accountVisual/accountVisual.css
@@ -1,4 +1,4 @@
-@import '../app/variables.css';
+@import '../../app/variables.css';
 
 .wrapper {
   display: inline-block;
