+++ src/components/toolbox/accountVisual/index.test.js
@@ -3,8 +3,8 @@ import { expect } from 'chai';
 import { shallow, mount } from 'enzyme';
 import sinon from 'sinon';
 import AccountVisual from './index';
-import accounts from '../../../test/constants/accounts';
-import breakpoints from '../../constants/breakpoints';
+import accounts from '../../../../test/constants/accounts';
+import breakpoints from '../../../constants/breakpoints';
 
 describe('AccountVisual', () => {
   it('should create account visual of an address', () => {
